+++ backend/handlers/comment.go
@@ -1,12 +1,45 @@
 package handlers
 
 import (
+	"backend/db"
 	"backend/events"
+	"backend/utils"
 	"encoding/json"
+	"errors"
 	"log"
 	"net/http"
+	"strconv"
+	"time"
 )
 
+func InsertComment(comment Comment) error {
+	id, err := db.InsertData("comments", comment.UserId, comment.PostId, comment.Content, "", time.Now())
+	if err != nil {
+		return errors.New("Error inserting comment " + err.Error())
+	}
+	if id == 0 {
+		return errors.New("error inserting comment ")
+	}
+	if comment.Image != "" {
+		// Process the image and save it to the local storage
+		str := strconv.Itoa(int(id))
+		url := "./images/comments/" + str
+		url, err := utils.ProcessImage(comment.Image, url)
+		if err != nil {
+			log.Println("Error processing comment image:", err)
+			//response = Response{err.Error(), events.Event{}, http.StatusBadRequest}
+			return err
+		}
+		comment.Image = url
+	} else {
+		comment.Image = ""
+	}
+	err = db.UpdateData("comments", comment.Image, id)
+	if err != nil {
+		return errors.New("Error updating comment image" + err.Error())
+	}
+	return nil
+}
 func CreateComment(payload json.RawMessage) (Response, error) {
 	var response Response
 	var comment Comment
